Refactor ClientLayout and add navigation components
- Replaced the existing layout structure in `ClientLayout.tsx` with `NavbarComponent` and `SidebarComponent` for improved organization.
- Moved theme toggle functionality to the new `NavbarComponent`.
- Updated `layout.tsx` to apply the Lato font globally.
- Removed unused CSS properties from `globals.css`.
