+++ apps/documentation/src/app/layout.tsx
@@ -1,7 +1,14 @@
 import type { Metadata } from "next";
+import { Lato } from "next/font/google";
 import "./globals.css";
 import ClientLayout from "./ClientLayout";
 import ThemeProvider from "./ThemeProvider";
+
+const lato = Lato({
+  subsets: ["latin"],
+  weight: ["300", "400", "700", "900"],
+  display: "swap",
+});
 
 export const metadata: Metadata = {
   title: "JustiFi Web Components Documentation",
@@ -11,7 +18,7 @@ export const metadata: Metadata = {
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en">
-      <body>
+      <body className={lato.className}>
         <ThemeProvider>
           <ClientLayout>{children}</ClientLayout>
         </ThemeProvider>
